+++ src/index.ts
@@ -19,7 +19,6 @@ extendEnvironment((hre: HardhatRuntimeEnvironment) => {
 
 export * from "./tasks";
 
-export * from "./bundler";
 export * from "./deployments";
 export * from "./docs";
 export * from "./flat";
