+++ src/tasks/index.ts
@@ -1,6 +1,7 @@
 export * from "./abi";
 export * from "./accounts";
 export * from "./blockscout";
+export * from "./aaBundler";
 export * from "./faucet";
 export * from "./uploadAbi";
 export * from "./mnemonic";
